backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.3
celery==5.5.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.8
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
kombu==5.5.4
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==5.2.1
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
